requirements.txt
streamlit
PyMuPDF
openai
pandas
tiktoken
scipy
PyYAML